Add ChannelProvider
