+++ src/mcp/java/net/minecraftforge/gradle/mcp/MCPChannelProvider.java
@@ -1,0 +1,42 @@
+/*
+ * ForgeGradle
+ * Copyright (C) 2018 Forge Development LLC
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+ * USA
+ */
+
+package net.minecraftforge.gradle.mcp;
+
+import net.minecraftforge.gradle.common.util.MavenArtifactDownloader;
+import org.gradle.api.Project;
+
+import javax.annotation.Nullable;
+import java.io.File;
+import java.io.IOException;
+
+public class MCPChannelProvider extends ChannelProvider {
+    public MCPChannelProvider() {
+        super("snapshot", "snapshot_nodoc", "stable", "stable_nodoc");
+    }
+
+    @Nullable
+    @Override
+    public File getMappingsFile(MCPRepo mcpRepo, Project project, String channel, String version) throws IOException {
+        String desc = "de.oceanlabs.mcp:mcp_" + channel + ":" + version + "@zip";
+        mcpRepo.debug("    Mapping: " + desc);
+        return MavenArtifactDownloader.manual(project, desc, false);
+    }
+}
