+++ src/mcp/java/net/minecraftforge/gradle/mcp/MCPRepo.java
@@ -20,21 +20,13 @@
 
 package net.minecraftforge.gradle.mcp;
 
-import com.google.common.base.Functions;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Lists;
-import com.google.gson.Gson;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonPrimitive;
+import com.google.common.collect.Maps;
 import net.minecraftforge.artifactural.api.artifact.ArtifactIdentifier;
 import net.minecraftforge.artifactural.api.repository.ArtifactProvider;
 import net.minecraftforge.artifactural.api.repository.Repository;
 import net.minecraftforge.artifactural.base.repository.ArtifactProviderBuilder;
 import net.minecraftforge.artifactural.base.repository.SimpleRepository;
 import net.minecraftforge.artifactural.gradle.GradleRepositoryAdapter;
-import net.minecraftforge.gradle.common.config.MCPConfigV2;
 import net.minecraftforge.gradle.common.util.BaseRepo;
 import net.minecraftforge.gradle.common.util.DownloadUtils;
 import net.minecraftforge.gradle.common.util.HashFunction;
@@ -49,43 +41,20 @@ import net.minecraftforge.gradle.common.util.VersionJson;
 import net.minecraftforge.gradle.mcp.util.MCPRuntime;
 import net.minecraftforge.gradle.mcp.util.MCPWrapper;
 import net.minecraftforge.srgutils.IMappingFile;
-import net.minecraftforge.srgutils.IMappingFile.INode;
-import net.minecraftforge.srgutils.IMappingFile.IPackage;
-import net.minecraftforge.srgutils.IMappingFile.IParameter;
-import net.minecraftforge.srgutils.IRenamer;
-import net.minecraftforge.srgutils.IMappingFile.IClass;
 import net.minecraftforge.srgutils.IMappingFile.IField;
 import net.minecraftforge.srgutils.IMappingFile.IMethod;
-
+import net.minecraftforge.srgutils.IRenamer;
 import org.apache.commons.io.FileUtils;
 import org.gradle.api.Project;
 import org.gradle.api.logging.Logger;
 
-import com.google.common.collect.Maps;
-import de.siegmar.fastcsv.writer.CsvWriter;
-import de.siegmar.fastcsv.writer.LineDelimiter;
+import javax.annotation.Nullable;
 import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 import java.net.URL;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.TreeMap;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
-import java.util.zip.ZipOutputStream;
-
-import javax.annotation.Nullable;
 
 /**
  * Provides the following artifacts:
@@ -123,8 +92,6 @@ public class MCPRepo extends BaseRepo {
     private static final String NAMES_MCP = "^(mcp_config)$";
     private static final String STEP_MERGE = "merge"; //TODO: Design better way to get steps output, for now hardcode
     private static final String STEP_RENAME = "rename";
-    private static final Pattern PARCHMENT_PATTERN = Pattern.compile("(?<mappingsversion>[\\w\\-.]+)-(?<mcpversion>(?<mcversion>[\\d.]+)(?:-\\d{8}\\.\\d{6})?)");
-    private static final Gson GSON = new Gson();
 
     //This is the artifact we expose that is a zip containing SRG->Official fields and methods.
     public static final String MAPPING_DEP = "net.minecraft:mappings_{CHANNEL}:{VERSION}@zip";
@@ -159,7 +126,12 @@ public class MCPRepo extends BaseRepo {
         return getInstance(project);
     }
 
-    private File cacheMC(String side, String version, @Nullable String classifier, String ext) {
+    @Override
+    protected File cache(String... path) {
+        return super.cache(path);
+    }
+
+    File cacheMC(String side, String version, @Nullable String classifier, String ext) {
         if (classifier != null)
             return cache("net", "minecraft", side, version, side + '-' + version + '-' + classifier + '.' + ext);
         return cache("net", "minecraft", side, version, side + '-' + version + '.' + ext);
@@ -172,10 +144,6 @@ public class MCPRepo extends BaseRepo {
     }
     private File cacheMCP(String version) {
         return cache("de", "oceanlabs", "mcp", "mcp_config", version);
-    }
-
-    private File cacheParchment(String mcpversion, String mappingsVersion, String ext) {
-        return cache("org", "parchmentmc", "data", "parchment-" + mcpversion, mappingsVersion, "parchment-" + mcpversion + '-' + mappingsVersion + '.' + ext);
     }
 
     @Override
@@ -233,13 +201,13 @@ public class MCPRepo extends BaseRepo {
         return null;
     }
 
-    private HashStore commonHash(File mcp) {
+    HashStore commonHash(File mcp) {
         return new HashStore(this.getCacheRoot())
             .add("mcp", mcp);
     }
 
     @Nullable
-    private File getMCP(String version) {
+    File getMCP(String version) {
         return MavenArtifactDownloader.manual(project, "de.oceanlabs.mcp:mcp_config:" + version + "@zip", false);
     }
 
@@ -372,7 +340,7 @@ public class MCPRepo extends BaseRepo {
     }
 
     @Nullable
-    private File findRenames(String classifier, IMappingFile.Format format, String version, boolean toObf) throws IOException {
+    File findRenames(String classifier, IMappingFile.Format format, String version, boolean toObf) throws IOException {
         String ext = format.name().toLowerCase();
         //File names = findNames(version));
         File mcp = getMCP(version);
@@ -402,15 +370,9 @@ public class MCPRepo extends BaseRepo {
         String channel = mapping.substring(0, idx);
         String version = mapping.substring(idx + 1);
 
-        if ("official".equals(channel)) {
-            return findOfficialMapping(version);
-        } else if ("parchment".equals(channel)) {
-            return findParchmentMapping(version);
-        } else if ("snapshot".equals(channel) || "snapshot_nodoc".equals(channel) || "stable".equals(channel) || "stable_nodoc".equals(channel)) { //MCP
-            String desc = "de.oceanlabs.mcp:mcp_" + channel + ":" + version + "@zip";
-            debug("    Mapping: " + desc);
-            return MavenArtifactDownloader.manual(project, desc, false);
-        }
+        ChannelProvider provider = ChannelProviders.getProvider(channel);
+        if (provider != null)
+            return provider.getMappingsFile(this, project, channel, version);
         throw new IllegalArgumentException("Unknown mapping provider: " + mapping);
     }
 
@@ -489,342 +451,6 @@ public class MCPRepo extends BaseRepo {
     }
 
     @Nullable
-    private File findOfficialMapping(String version) throws IOException {
-        String mcpversion = version;
-        int idx = version.lastIndexOf('-');
-        if (idx != -1 && MinecraftRepo.MCP_CONFIG_VERSION.matcher(version.substring(idx + 1)).matches()) {
-            //mcpversion = version.substring(idx);
-            version = version.substring(0, idx);
-        }
-        File client = MavenArtifactDownloader.generate(project, "net.minecraft:client:" + version + ":mappings@txt", true);
-        File server = MavenArtifactDownloader.generate(project, "net.minecraft:server:" + version + ":mappings@txt", true);
-        if (client == null || server == null)
-            throw new IllegalStateException("Could not create " + mcpversion + " official mappings due to missing ProGuard mappings.");
-
-        File tsrg = findRenames("obf_to_srg", IMappingFile.Format.TSRG, mcpversion, false);
-        if (tsrg == null)
-            throw new IllegalStateException("Could not create " + mcpversion + " official mappings due to missing MCP's tsrg");
-
-        File mcp = getMCP(mcpversion);
-        if (mcp == null)
-            return null;
-
-        File mappings = cacheMC("mapping", mcpversion, "mapping", "zip");
-        HashStore cache = commonHash(mcp)
-                .load( cacheMC("mapping", mcpversion, "mapping", "zip.input"))
-                .add("pg_client", client)
-                .add("pg_server", server)
-                .add("tsrg", tsrg)
-                .add("codever", "2");
-
-        if (!cache.isSame() || !mappings.exists()) {
-            IMappingFile pg_client = IMappingFile.load(client);
-            IMappingFile pg_server = IMappingFile.load(server);
-
-            //Verify that the PG files merge, merge in MCPConfig, but doesn't hurt to double check here.
-            //And if we don't we need to write a handler to spit out correctly sided info.
-
-            IMappingFile srg = IMappingFile.load(tsrg);
-
-            Map<String, String> cfields = new TreeMap<>();
-            Map<String, String> sfields = new TreeMap<>();
-            Map<String, String> cmethods = new TreeMap<>();
-            Map<String, String> smethods = new TreeMap<>();
-
-            for (IClass cls : pg_client.getClasses()) {
-                IClass obf = srg.getClass(cls.getMapped());
-                if (obf == null) // Class exists in official source, but doesn't make it past obfusication so it's not in our mappings.
-                    continue;
-                for (IField fld : cls.getFields()) {
-                    String name = obf.remapField(fld.getMapped());
-                    if (name.startsWith("field_") || name.startsWith("f_"))
-                        cfields.put(name, fld.getOriginal());
-                }
-                for (IMethod mtd : cls.getMethods()) {
-                    String name = obf.remapMethod(mtd.getMapped(), mtd.getMappedDescriptor());
-                    if (name.startsWith("func_") || name.startsWith("m_"))
-                        cmethods.put(name, mtd.getOriginal());
-                }
-            }
-            for (IClass cls : pg_server.getClasses()) {
-                IClass obf = srg.getClass(cls.getMapped());
-                if (obf == null) // Class exists in official source, but doesn't make it past obfusication so it's not in our mappings.
-                    continue;
-                for (IField fld : cls.getFields()) {
-                    String name = obf.remapField(fld.getMapped());
-                    if (name.startsWith("field_") || name.startsWith("f_"))
-                        sfields.put(name, fld.getOriginal());
-                }
-                for (IMethod mtd : cls.getMethods()) {
-                    String name = obf.remapMethod(mtd.getMapped(), mtd.getMappedDescriptor());
-                    if (name.startsWith("func_") || name.startsWith("m_"))
-                        smethods.put(name, mtd.getOriginal());
-                }
-            }
-
-            String[] header = new String[] {"searge", "name", "side", "desc"};
-            List<String[]> fields = new ArrayList<>();
-            List<String[]> methods = new ArrayList<>();
-            fields.add(header);
-            methods.add(header);
-
-            for (String name : cfields.keySet()) {
-                String cname = cfields.get(name);
-                String sname = sfields.get(name);
-                if (cname.equals(sname)) {
-                    fields.add(new String[]{name, cname, "2", ""});
-                    sfields.remove(name);
-                } else
-                    fields.add(new String[]{name, cname, "0", ""});
-            }
-
-            for (String name : cmethods.keySet()) {
-                String cname = cmethods.get(name);
-                String sname = smethods.get(name);
-                if (cname.equals(sname)) {
-                    methods.add(new String[]{name, cname, "2", ""});
-                    smethods.remove(name);
-                } else
-                    methods.add(new String[]{name, cname, "0", ""});
-            }
-
-            sfields.forEach((k,v) -> fields.add(new String[] {k, v, "1", ""}));
-            smethods.forEach((k,v) -> methods.add(new String[] {k, v, "1", ""}));
-
-            if (!mappings.getParentFile().exists())
-                mappings.getParentFile().mkdirs();
-
-            try (FileOutputStream fos = new FileOutputStream(mappings);
-                 ZipOutputStream out = new ZipOutputStream(fos)) {
-                writeCsv("fields.csv", fields, out);
-                writeCsv("methods.csv", methods, out);
-            }
-
-            cache.save();
-            Utils.updateHash(mappings, HashFunction.SHA1);
-        }
-
-
-        return mappings;
-    }
-
-    private File findParchmentMapping(String version) throws IOException {
-        // Format is {MAPPINGS_VERSION}-{MC_VERSION}-{MCP_VERSION} where MCP_VERSION is optional
-        Matcher matcher = PARCHMENT_PATTERN.matcher(version);
-        if (!matcher.matches())
-            throw new IllegalStateException("Parchment version of " + version + " is invalid");
-
-        String mappingsversion = matcher.group("mappingsversion");
-        String mcversion = matcher.group("mcversion");
-        String mcpversion = matcher.group("mcpversion");
-
-        File client = MavenArtifactDownloader.generate(project, "net.minecraft:client:" + mcversion + ":mappings@txt", true);
-        if (client == null)
-            throw new IllegalStateException("Could not create " + mcversion + " official mappings due to missing ProGuard mappings.");
-
-        File tsrg = findRenames("obf_to_srg", IMappingFile.Format.TSRG2, mcpversion, false);
-        if (tsrg == null)
-            throw new IllegalStateException("Could not create " + mcpversion + " parchment mappings due to missing MCP's tsrg");
-
-        File mcp = getMCP(mcpversion);
-        if (mcp == null)
-            return null;
-
-        String artifact = "org.parchmentmc.data:parchment-" + mcversion + ":" + mappingsversion + ":checked@zip";
-        File dep = MavenArtifactDownloader.manual(project, artifact, false);
-        if (dep == null) {
-            // TODO remove this later? or keep backwards-compatibility with older releases?
-            dep = MavenArtifactDownloader.manual(project, artifact.replace(":checked", ""), false);
-        }
-        if (dep == null)
-            throw new IllegalStateException("Could not find Parchment version of " + mappingsversion + '-' + mcversion + " with artifact " + artifact);
-
-        File mappings = cacheParchment(mcpversion, mappingsversion, "zip");
-        HashStore cache = commonHash(mcp)
-                .load(cacheParchment(mcpversion, mappingsversion, "zip.input"))
-                .add("mcversion", version)
-                .add("mappings", dep)
-                .add("tsrg", tsrg)
-                .add("codever", "1");
-
-        if (cache.isSame() && mappings.exists())
-            return mappings;
-
-        boolean official = MCPConfigV2.getFromArchive(mcp).isOfficial();
-        try (ZipFile zip = new ZipFile(dep)) {
-            ZipEntry entry = zip.getEntry("parchment.json");
-            if (entry == null)
-                throw new IllegalStateException("Parchment zip did not contain \"parchment.json\"");
-
-            JsonObject json = GSON.fromJson(new InputStreamReader(zip.getInputStream(entry)), JsonObject.class);
-            String specversion = json.get("version").getAsString();
-            if (!specversion.startsWith("1."))
-                throw new IllegalStateException("Parchment mappings spec version was " + specversion + " and did not start with \"1.\", cannot parse!");
-            IMappingFile srg = IMappingFile.load(tsrg);
-            IMappingFile mojToObf = IMappingFile.load(client);
-            // Have to do it this way to preserve parameters and eliminate SRG classnames
-            IMappingFile mojToSrg = srg.reverse().chain(mojToObf.reverse()).reverse().rename(new IRenamer() {
-                @Override
-                public String rename(IClass value) {
-                    return value.getOriginal();
-                }
-            });
-
-            String[] header = {"searge", "name", "desc"};
-            List<String[]> packages = Lists.<String[]>newArrayList(header);
-            List<String[]> classes = Lists.<String[]>newArrayList(header);
-            List<String[]> fields = Lists.<String[]>newArrayList(header);
-            List<String[]> methods = Lists.<String[]>newArrayList(header);
-            List<String[]> parameters = Lists.<String[]>newArrayList(header);
-
-            Map<String, JsonObject> classMap = getNamedJsonMap(json.getAsJsonArray("classes"), false);
-            Map<String, JsonObject> packageMap = getNamedJsonMap(json.getAsJsonArray("packages"), false);
-
-            for (IPackage srgPackage : mojToSrg.getPackages()) {
-                JsonObject pckg = packageMap.get(srgPackage.getOriginal());
-                populateMappings(packages, null, srgPackage, pckg);
-            }
-            for (IClass srgClass : mojToSrg.getClasses()) {
-                JsonObject cls = classMap.get(srgClass.getOriginal());
-                populateMappings(classes, srgClass, srgClass, cls);
-
-                Map<String, JsonObject> fieldMap = cls == null ? ImmutableMap.of() : getNamedJsonMap(cls.getAsJsonArray("fields"), false);
-                for (IField srgField : srgClass.getFields()) {
-                    populateMappings(fields, srgClass, srgField, fieldMap.get(srgField.getOriginal()));
-                }
-
-                Map<String, JsonObject> methodMap = cls == null ? ImmutableMap.of() : getNamedJsonMap(cls.getAsJsonArray("methods"), true);
-                for (IMethod srgMethod : srgClass.getMethods()) {
-                    JsonObject method = methodMap.get(srgMethod.getOriginal() + srgMethod.getDescriptor());
-                    StringBuilder mdJavadoc = new StringBuilder(getJavadocs(method));
-                    List<IParameter> srgParams = new ArrayList<>(srgMethod.getParameters());
-                    if (method != null && method.has("parameters")) {
-                        JsonArray jsonParams = method.getAsJsonArray("parameters");
-                        if (!official || jsonParams.size() == srgParams.size())
-                            for (int i = 0; i < jsonParams.size(); i++) {
-                                JsonObject parameter = jsonParams.get(i).getAsJsonObject();
-                                boolean isConstructor = method.get("name").getAsString().equals("<init>");
-                                String srgParam;
-                                if (official) {
-                                    srgParam = srgParams.get(i).getMapped();
-                                } else {
-                                    String srgId = srgMethod.getMapped().indexOf('_') == -1
-                                            ? srgMethod.getMapped()
-                                            : srgMethod.getMapped().split("_")[1];
-                                    srgParam = String.format(isConstructor ? "p_i%s_%s_" : "p_%s_%s_", srgId, parameter.get("index").getAsString());
-                                }
-                                String paramName = parameter.has("name") ? parameter.get("name").getAsString() : null;
-                                if (paramName != null) {
-                                    parameters.add(new String[]{srgParam, paramName, ""});
-                                }
-                                String paramJavadoc = getJavadocs(parameter);
-                                if (!paramJavadoc.isEmpty())
-                                    mdJavadoc.append("\\n@param ").append(paramName != null ? paramName : srgParam).append(' ').append(paramJavadoc);
-                            }
-                    }
-                    populateMappings(methods, srgClass, srgMethod, mdJavadoc.toString());
-                }
-            }
-
-            if (!mappings.getParentFile().exists())
-                mappings.getParentFile().mkdirs();
-
-            try (FileOutputStream fos = new FileOutputStream(mappings);
-                    ZipOutputStream out = new ZipOutputStream(fos)) {
-                writeCsv("classes.csv", classes, out);
-                writeCsv("fields.csv", fields, out);
-                writeCsv("methods.csv", methods, out);
-                writeCsv("params.csv", parameters, out);
-                writeCsv("packages.csv", packages, out);
-            }
-        }
-
-        cache.save();
-        Utils.updateHash(mappings, HashFunction.SHA1);
-
-        return mappings;
-    }
-
-    private Map<String, JsonObject> getNamedJsonMap(JsonArray array, boolean hasDescriptor) {
-        if (array == null || array.size() == 0)
-            return ImmutableMap.of();
-        return StreamSupport.stream(array.spliterator(), false)
-                .map(JsonObject.class::cast)
-                .collect(Collectors.toMap(j -> {
-                    String key = j.get("name").getAsString();
-                    if (hasDescriptor)
-                        key += j.get("descriptor").getAsString();
-                    return key;
-                }, Functions.identity()));
-    }
-
-    private void populateMappings(List<String[]> mappings, IClass srgClass, INode srgNode, JsonObject json) {
-        populateMappings(mappings, srgClass, srgNode, getJavadocs(json));
-    }
-
-    private void populateMappings(List<String[]> mappings, IClass srgClass, INode srgNode, String desc) {
-        if (srgNode instanceof IPackage || srgNode instanceof IClass) {
-            String name = srgNode.getMapped().replace('/', '.');
-            // TODO fix InstallerTools so that we don't have to expand the csv size for no reason
-            if (!desc.isEmpty())
-                mappings.add(new String[]{name, name, desc});
-            return;
-        }
-        String srgName = srgNode.getMapped();
-        String mojName = srgNode.getOriginal();
-        boolean isSrg = srgName.startsWith("p_") || srgName.startsWith("func_") || srgName.startsWith("m_") || srgName.startsWith("field_") || srgName.startsWith("f_");
-        // If it's not a srg id and has javadocs, we need to add the class to the beginning as it is a special method/field of some kind
-        if (!isSrg && !desc.isEmpty() && (srgNode instanceof IMethod || srgNode instanceof IField)) {
-            srgName = srgClass.getMapped().replace('/', '.') + '#' + srgName;
-        }
-        // Only add to the mappings list if it is mapped or has javadocs
-        if ((isSrg && !srgName.equals(mojName)) || !desc.isEmpty())
-            mappings.add(new String[]{srgName, mojName, desc});
-    }
-
-    private String getJavadocs(JsonObject json) {
-        if (json == null)
-            return "";
-        JsonElement element = json.get("javadoc");
-        if (element == null)
-            return "";
-        if (element instanceof JsonPrimitive)
-            return element.getAsString(); // Parameters don't use an array for some reason
-        if (!(element instanceof JsonArray))
-            return "";
-        JsonArray array = (JsonArray) element;
-        StringBuilder sb = new StringBuilder();
-        int size = array.size();
-        for (int i = 0; i < size; i++) {
-            sb.append(array.get(i).getAsString());
-            if (i != size - 1)
-                sb.append("\\n");
-        }
-        return sb.toString();
-    }
-
-    private void writeCsv(String name, List<String[]> mappings, ZipOutputStream out) throws IOException {
-        if (mappings.size() <= 1)
-            return;
-        out.putNextEntry(Utils.getStableEntry(name));
-        try (CsvWriter writer = CsvWriter.builder().lineDelimiter(LineDelimiter.LF).build(new UncloseableOutputStreamWriter(out))) {
-            mappings.forEach(writer::writeRow);
-        }
-        out.closeEntry();
-    }
-
-    private class UncloseableOutputStreamWriter extends OutputStreamWriter {
-        public UncloseableOutputStreamWriter(OutputStream out) {
-            super(out);
-        }
-
-        @Override
-        public void close() throws IOException {
-            super.flush();
-        }
-    }
-
-    @Nullable
     private File findEmptyPom(String side, String version) throws IOException {
         File pom = cacheMC(side, version, null, "pom");
         debug("    Finding pom: " + pom);
@@ -840,5 +466,10 @@ public class MCPRepo extends BaseRepo {
         }
 
         return pom;
+    }
+
+    @Override
+    protected void debug(String message) {
+        super.debug(message);
     }
 }
